Combined the auth tests into one module
