+++ tests/test_auth.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Test cases for the token handling functionality in firebase_admin.auth module."""
+"""Test cases for the firebase_admin.auth module."""
 import os
 import time
 
@@ -38,6 +38,9 @@ MOCK_CREDENTIAL = credentials.Certificate(
 MOCK_PUBLIC_CERTS = testutils.resource('public_certs.json')
 MOCK_PRIVATE_KEY = testutils.resource('private_key.pem')
 MOCK_SERVICE_ACCOUNT_EMAIL = MOCK_CREDENTIAL.service_account_email
+
+INVALID_STRINGS = [0, 1, True, False, list(), tuple(), dict()]
+INVALID_BOOLS = [None, '', 'foo', 0, 1, list(), tuple(), dict()]
 
 
 class AuthFixture(object):
@@ -256,3 +259,115 @@ class TestVerifyIdToken(object):
         auth._request = testutils.MockRequest(404, 'not found')
         with pytest.raises(exceptions.TransportError):
             authtest.verify_id_token(TEST_ID_TOKEN)
+
+
+@pytest.fixture(scope='module')
+def user_mgt_app():
+    app = firebase_admin.initialize_app(testutils.MockCredential(), name='userMgt')
+    yield app
+    firebase_admin.delete_app(app)
+
+def _instrument_user_manager(app, status, payload):
+    auth_service = auth._get_auth_service(app)
+    user_manager = auth_service.user_manager
+    user_manager._session.mount(
+        auth._UserManager._ID_TOOLKIT_URL,
+        testutils.MockAdapter(payload, status, []))
+    return user_manager
+
+def _check_user_record(user):
+    assert user.uid == 'testuser'
+    assert user.email == 'testuser@example.com'
+    assert user.display_name == 'Test User'
+    assert user.photo_url == 'http://www.example.com/testuser/photo.png'
+    assert user.disabled is False
+    assert user.email_verified is True
+    assert user.user_metadata.creation_timestamp == 1234567890
+    assert user.user_metadata.last_sign_in_timestamp is None
+    assert user.provider_id == 'firebase'
+    assert len(user.provider_data) == 1
+    provider = user.provider_data[0]
+    assert provider.uid == 'testuser@example.com'
+    assert provider.email == 'testuser@example.com'
+    assert provider.display_name == 'Test User'
+    assert provider.photo_url == 'http://www.example.com/testuser/photo.png'
+    assert provider.provider_id == 'password'
+
+
+class TestGetUser(object):
+
+    @pytest.mark.parametrize('arg', INVALID_STRINGS)
+    def test_invalid_get_user(self, arg):
+        with pytest.raises(ValueError):
+            auth.get_user(arg)
+
+    def test_get_user(self, user_mgt_app):
+        _instrument_user_manager(user_mgt_app, 200, testutils.resource('get_user.json'))
+        _check_user_record(auth.get_user('testuser', user_mgt_app))
+
+    @pytest.mark.parametrize('arg', INVALID_STRINGS)
+    def test_invalid_get_user_by_email(self, arg):
+        with pytest.raises(ValueError):
+            auth.get_user(arg)
+
+    def test_get_user_by_email(self, user_mgt_app):
+        _instrument_user_manager(user_mgt_app, 200, testutils.resource('get_user.json'))
+        _check_user_record(auth.get_user_by_email('testuser@example.com', user_mgt_app))
+
+    def test_get_user_non_existing(self, user_mgt_app):
+        _instrument_user_manager(user_mgt_app, 200, '{"users":[]}')
+        with pytest.raises(auth.FirebaseAuthError) as excinfo:
+            auth.get_user('testuser', user_mgt_app)
+        assert excinfo.value.code == 'USER_NOT_FOUND_ERROR'
+
+    def test_get_user_http_error(self, user_mgt_app):
+        _instrument_user_manager(user_mgt_app, 500, '{}')
+        with pytest.raises(auth.FirebaseAuthError) as excinfo:
+            auth.get_user('testuser', user_mgt_app)
+        assert excinfo.value.code == 'INTERNAL_ERROR'
+
+
+class TestCreateUser(object):
+
+    @pytest.mark.parametrize('arg', INVALID_STRINGS + ['a'*129])
+    def test_invalid_uid(self, arg):
+        with pytest.raises(ValueError):
+            auth.create_user({'uid' : arg})
+
+    @pytest.mark.parametrize('arg', INVALID_STRINGS + ['not-an-email'])
+    def test_invalid_email(self, arg):
+        with pytest.raises(ValueError):
+            auth.create_user({'email' : arg})
+
+    @pytest.mark.parametrize('arg', INVALID_STRINGS)
+    def test_invalid_display_name(self, arg):
+        with pytest.raises(ValueError):
+            auth.create_user({'displayName' : arg})
+
+    @pytest.mark.parametrize('arg', INVALID_STRINGS + ['not-a-url'])
+    def test_invalid_photo_url(self, arg):
+        with pytest.raises(ValueError):
+            auth.create_user({'photoUrl' : arg})
+
+    @pytest.mark.parametrize('arg', INVALID_STRINGS + ['short'])
+    def test_invalid_password(self, arg):
+        with pytest.raises(ValueError):
+            auth.create_user({'password' : arg})
+
+    @pytest.mark.parametrize('arg', INVALID_BOOLS)
+    def test_invalid_email_verified(self, arg):
+        with pytest.raises(ValueError):
+            auth.create_user({'emailVerified' : arg})
+
+    @pytest.mark.parametrize('arg', INVALID_BOOLS)
+    def test_invalid_disabled(self, arg):
+        with pytest.raises(ValueError):
+            auth.create_user({'disabled' : arg})
+
+    def test_invalid_property(self):
+        with pytest.raises(ValueError):
+            auth.create_user({'unsupported' : 'value'})
+
+    def test_create_user(self, user_mgt_app):
+        user_mgt = _instrument_user_manager(user_mgt_app, 200, '{"localId":"testuser"}')
+        assert user_mgt.create_user() == 'testuser'
